+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
   - Use below extensions on `context` to easily navigate between screens
 
-  [1] `navTo()` [works as `Navigator.push()`]
+  [1] `goTo()` [works as `Navigator.push()`]
 
-  [2] `navRplcmntTo()` [works as `Navigatoe.pushReplacement()`]
+  [2] `rplcmntgoTo()` [works as `Navigatoe.pushReplacement()`]
 
   [3] `pop()` [works as `Navigator.pop()`]
+
+  [4] `goToNamed` [works as `Navigator.pushNamed()`]
 
 - New `Shaper` class added in the creators.
   - Helps in shaping widgets which contains `shape` parameter.